communication
intention clashing fixed.
Note: intentions can still clash, because we don't communicate our
intentions all the time but at every x number of steps (which is
changeable based on the environment). This is because we can't send
more than 3 things at a time, so we have to choose wisely if there's a
point in communicating current intention or sending information. This
could be represented as a GA parameter.